+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The Union method on a IEnumerable<Range<T>> to union two sequences and also reduce the result
 - The Intersect method on a IEnumerable<Range<T>>, returns a new IEnumerable<Range<T>> where only ranges that occur in both sequences are returned
 - IsCompletelyBefore and IsCompletelyBehind methods on Range<T>, the return if a given Range<T> is lower than or greater than the specified range and that they do not touch nor overlap each other
+- The Inverse method on IEnumerable<Range<T>>, returns a new IEnumerable<Range<T>> with all gaps between the given ranges and also a range from the minvalue to the start of the first range and the end of the last range to the maxvalue
 
 ### Removed 
 - RangeCollection<T> has been removed in favor of Linq extension methods on IEnumerable<Range<T>>